Xpath queries to use internal xpath query function
Posh Core and 'full' (desktop) have different classes for
performing xpath queries on xml.

This commit implements an internal function that leverages
the appropriate method for the edition its running on.

All xpath query references now point to it rather than the native
method

This is a seriously breaking change. :)