Improve error handling in invoke-nsxresmethod/webrequest
In order to support core and full versions of PowerShell with the
one module, some better handling of the differing exception
types thrown when invoke-rest/web is called was needed.

This change allows the user to receive any error text sent back
from the API if it was received, or to get the underlying EX if not.