Update New-NsxController with job status checking
Previously, when we did a new controller api request,
assuming the request was accepted, we blindly assumed the controller then existed.

Creating a controller is an async task, and may fail (out
of storage, vc failed for some reason etc) and cause the
controller to be removed.

When we subsequently tried to check for it, getting a
non existant controller would fail.

Previous partial attempts at handling this focussed on
catching the various failure states in the get for a
missing controller, but this change actually monitors
the state of the deployment job using the nsx api to
determine success (and before the new controller is
actually retreived)

Net result - controller deployment failure is now
handled gracefullly with a meaningful error returned to
the user instead of some internal failure of PowerNSX
code.